Moved find_bounding_box to utils
