+++ openmc_dagmc_wrapper/__init__.py
@@ -1,7 +1,7 @@
-from .utils import process_results, create_material
+from .utils import process_results, create_material, find_bounding_box
 
 from .Geometry import Geometry
 from .Materials import Materials
 from .Tally import CellTally, CellTallies, MeshTally2D, MeshTallies2D, \
-    MeshTally3D, MeshTallies3D, find_bounding_box
+    MeshTally3D, MeshTallies3D
 from .Settings import FusionSettings
